universities/pyproject.toml
[tool.poetry]
name = "universities"
version = "0.1.0"
description = ""
authors = ["iizhukov <iizhukov@iicloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4"
django-rest-framework = "^0.1.0"
django-cors-headers = "^4.3.1"
cryptography = "^42.0.5"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
pyjwt = "^2.8.0"
pillow = "^10.3.0"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
runserver = "scripts.runserver:main"
